// src/interfaces/product.interface.ts
export interface Product {
  id: string;
  description: string;
  images: string[];
  inStock: number;
  price: number;
  sizes: Size[];
  slug: string;
  tags: string[];
  title: string;
  // type: Type;
  gender: Category;
}

export interface CartProducts {
  id: string;
  slug: string;
  title: string;
  price: number;
  quantity: number;
  size: Size;
  image: string;
}

export interface ProductImage {
  id: number;
  url: string;
  productId?: string;
}

type Category = "men" | "women" | "kid" | "unisex";
export type Size = "XS" | "S" | "M" | "L" | "XL" | "XXL" | "XXXL";
export type Type = "shirts" | "pants" | "hoodies" | "hats";

// export interface RequestPagination {
//   currentPage: number;
//   totalPages: number;
//   products: {
//     images: string[];
//     ProductImage: { url: string }[];
//     id: string;
//     title: string;
//     description: string;
//     inStock: number;
//     price: number;
//     sizes: Size[];
//     slug: string;
//     tags: string[];
//     gender: Category;
//     categoryId: string;
//   };
// }